// src/features/auth/forgot-password/api/types.ts
// Типы для восстановления пароля

// Тип для ПЕРВОЙ отправки письма восстановления (эндпоинт: /auth/password-recovery)
export type SendRecoveryEmailType = {
  email: string // Email пользователя для отправки письма
  recaptcha: string
}

// Тип для ПОВТОРНОЙ отправки письма (эндпоинт: /auth/password-recovery-resending)
export type ResendRecoveryEmailType = {
  email: string // Email пользователя для повторной отправки
}

// Тип для установки НОВОГО ПАРОЛЯ (эндпоинт: /auth/new-password)
export type SetNewPasswordType = {
  newPassword: string // Новый пароль пользователя
  recoveryCode: string // Код восстановления из письма
}